+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -8,6 +8,8 @@ CHANGELOG
  * Add `$exclude` to `tagged_iterator` and `tagged_locator` configurator
  * Add an `env` function to the expression language provider
  * Add an `Autowire` attribute to tell a parameter how to be autowired
+ * Allow using expressions as service factories
+ * Deprecate `ReferenceSetArgumentTrait`
 
 6.0
 ---